dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev-requirements.in
#
appdirs==1.4.4            # via black, virtualenv
appnope==0.1.0            # via ipython
asgiref==3.3.1            # via -c requirements.txt, -r requirements.txt, django
attrs==20.3.0             # via -c requirements.txt, -r requirements.txt, cattrs
backcall==0.2.0           # via ipython
beautifulsoup4==4.9.3     # via -c requirements.txt, -r requirements.txt, django-bootstrap4
black==20.8b1             # via -r dev-requirements.in
cattrs==1.1.1             # via -c requirements.txt, -r requirements.txt
cfgv==3.2.0               # via pre-commit
click==7.1.2              # via black
coverage==5.3             # via coverage
decorator==4.4.2          # via ipython
distlib==0.3.1            # via virtualenv
django-bootstrap-datepicker-plus==3.0.5  # via -c requirements.txt, -r requirements.txt
django-bootstrap4==2.3.1  # via -c requirements.txt, -r requirements.txt
django-debug-toolbar==3.1.1  # via -r dev-requirements.in
django==3.1.3             # via -c requirements.txt, -r requirements.txt, django-bootstrap-datepicker-plus, django-bootstrap4, django-debug-toolbar, djangorestframework, djangorestframework-simplejwt
djangorestframework-simplejwt==4.5.0  # via -c requirements.txt, -r requirements.txt
djangorestframework==3.12.2  # via -c requirements.txt, -r requirements.txt, djangorestframework-simplejwt
filelock==3.0.12          # via virtualenv
identify==1.5.9           # via pre-commit
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via -r dev-requirements.in
isort==5.6.4              # via -r dev-requirements.in
jedi==0.17.2              # via ipython
mypy-extensions==0.4.3    # via black
nodeenv==1.5.0            # via pre-commit
parso==0.7.1              # via jedi
pathspec==0.8.1           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==8.0.1             # via -c requirements.txt, -r requirements.txt
pre-commit==2.8.2         # via -r dev-requirements.in
prompt-toolkit==3.0.8     # via ipython
ptyprocess==0.6.0         # via pexpect
pydocstyle==5.1.1         # via -r dev-requirements.in
pygments==2.7.2           # via ipython
pyjwt==1.7.1              # via -c requirements.txt, -r requirements.txt, djangorestframework-simplejwt
pytz==2020.4              # via -c requirements.txt, -r requirements.txt, django
pyyaml==5.3.1             # via pre-commit
regex==2020.10.28         # via black
six==1.15.0               # via virtualenv
snowballstemmer==2.0.0    # via pydocstyle
soupsieve==2.0.1          # via -c requirements.txt, -r requirements.txt, beautifulsoup4
sqlparse==0.4.1           # via -c requirements.txt, -r requirements.txt, django, django-debug-toolbar
toml==0.10.2              # via black, pre-commit
traitlets==5.0.5          # via ipython
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
virtualenv==20.1.0        # via pre-commit
wcwidth==0.2.5            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
